compatibility with imagej version 1.45s
removed some function calls that were not implemented in older versions of imagej
ij.jar library changed to 1.45s so using methods from newer versions cause compilation errors

Compatibility with older versions is problematic because I believe they do not support  different overlay in each stack slice. We need that for showing detections in the original image. Maybe we could write it so that when older version of imagej is used, some features won't work but the program will run succesfully.

fixes issue #68