shortened and removed colons from created image names
done because of issue #120
The issue is a bug in JRE 1.7(maybe also other versions) in java.awt.FileDialog. It mangles the default filename and user input together. If the mangled filename contains colon, it hangs or fails to write the file, otherwise only the filename is wrong. The problem only appears when the user selected filename is shorter than the default, so i made it a little shorter. Does not fix completely the problem, but the consequences are not as severe.

The bug is fixed in JRE 8